+++ pkg/util/refs.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -77,6 +77,17 @@ func WithDefault[T interface{}](in *T) T {
 	}
 
 	return *in
+}
+
+// IsDefault returns if object is default
+func IsDefault[T any](in T) bool {
+	t := reflect.ValueOf(in)
+
+	if t.IsValid() {
+		return false
+	}
+
+	return t.IsZero() || t.IsNil()
 }
 
 // Default returns generic default value for type T